requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
pytesseract>=0.3.10
streamlit>=1.28.0
plotly>=5.17.0
openpyxl>=3.1.0

# Note: You also need to install Tesseract OCR separately
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Mac: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr
